switch to rust 2021 edition, bump MSRV to 1.56
use the latest edition to ensure that we get the linter warnings (there
were none, though) and other new features (though we currently don't
need them). this increases the MSRV to 1.56 as this is the first version
supporting the 2021 edition.

if a consumer cannot upgrade their rust version they can still continue
using the previous release of the library - there was no functional
change since then and they both support the same `oracle` & `r2d2`
releases.