ignore RUSTSEC-2020-0071 for chrono/time in `cargo audit` in CI
`chrono` pulls in `time:0.1` which has a known vulnerability which is
fixed in newer versions of `time`. however, `chrono` doesn't update to
that newer version...

acc. to chronotope/chrono#602 `chrono` isn't using the vulnerable API of
`time` and is thus safe to use (the same bug existed in `chrono` but has
been fixed there).

hopefully `chrono` will eventually update to a fixed version of `time`
as this seems to have caused major fallout in the ecosystem based on the
references to the tickets and comments on it.

note that we're not using `chrono` directly - it comes in through
`oracle`.