CI: use checked-in `Cargo.lock` for MSRV test
the transient dependency `log` requires a newer MSRV starting with
0.4.18. this way we force it to 0.4.17 which is compatible with our
MSRV. the same is the case for `chrono` which changed with 0.4.27 and
which we accordingly force to 0.4.26 for the MSRV check.
this is only used with the MSRV check, for all other tests we let
cargo test against the latest version.

note also that the cargo team has recently published an article on their
[change in guidance on commiting lockfiles][blog] which also suggests
providing a `Cargo.lock` for the MSRV build.

[blog]: https://blog.rust-lang.org/2023/08/29/committing-lockfiles.html